// src/app/model/dbtest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Dbtest { 
    dbtypesys1?: string;
    endtssys1?: string;
    jdbcurlsys1?: string;
    passwordsys1?: string;
    starttssys1?: string;
    tablenamesys1?: string;
    usernamesys1?: string;
}